pyproject.toml
[project]
name = "redfive"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "jinja2>=3.1.6",
    "jsonschema>=4.25.1",
    "networkx>=3.5",
    "openai>=2.5.0",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.1.1",
    "pyyaml>=6.0.3",
    "sqlalchemy>=2.0.44",
]
